Move components and hooks for search to ui-components
