+++ scm-ui/ui-components/src/search/HighlightedFragment.tsx
