+++ scm-ui/ui-components/src/search/Hit.tsx
